// Movie on Tip/client/src/model/IMovieList.tsx
export default interface IMovieList {
    id: number
    title: string
    year: number
    genres: []
    ratings: []
    poster: string
    contentRating: string
    duration: string
    releaseDate: string
    averageRating: number
    originalTitle: string
    storyline: string
    actors: []
    imdbRating: number
    posterurl: string
}